Use env-only SMTP advertised hostname in tests and service
