+++ apps/server/Services/AliasVault.SmtpService/appsettings.json
@@ -1,7 +1,4 @@
 {
-    "SmtpService": {
-        "AdvertisedHostname": ""
-    },
     "Logging": {
         "LogLevel": {
             "Default": "Information",
